+++ changelog.md
@@ -18,6 +18,9 @@
 * Added Pathfinding feature
   * Mobs are able to pathfind towards the target even if there's no block directly below them
   * Fixes mobs walking over open trapdoors
+* Changed Drowned Swim Speed Attribute feature
+  * Now Swimmers feature under Mobs module
+  * Added fish swim speed config option (defaults to 2.5x vanilla speed)
 * Only 50% of mobs will get the xray range override
 * Fisher Mobs
   * Fishing range is now configurable